TableComplexityIndexer/extraction/requirements_dots.txt

torch
torchvision
transformers
tokenizers
accelerate

qwen_vl_utils
modelscope
huggingface-hub
pillow
scikit-learn
beautifulsoup4
tqdm
pylatexenc
apted
lxml
